LRIS-32474 Move sidebar related components into separate component
