+++ src/components/molecules/Accordion/styles.module.scss
