+++ src/components/molecules/index.js
@@ -1,0 +1,3 @@
+import Accordion from './Accordion'
+
+export { Accordion }
