+++ src/components/organisms/Sidebar/styles.module.scss
@@ -1,0 +1,3 @@
+.sidebar {
+  padding: 5rem 3rem;
+}
