+++ src/components/templates/Blueprints/styles.module.scss
@@ -1,7 +1,3 @@
 .contentContainer {
   margin-left: 350px;
 }
-
-.sidebar {
-  padding: 5rem 3rem;
-}
